requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
datasets>=2.14.0        # HuggingFace datasets (streaming)
matplotlib>=3.7.0       # Training curve plots
tqdm>=4.65.0            # Progress bars
opencv-python>=4.8.0    # Optional: additional image processing
